Default to a local watermark when we are offline (#4051)
Improves the watermarking:
- We now have a `watermark.png` asset in our assets folder. When
releasing a new package this file now also gets uploaded to our cdn,
just like our static assets.
- We also have a local `watermark.png` image, which we use when using
our packages in a offline scenario. This file is now part of our package
and it gets copied from the assets folder when `refresh-assets` runs.

This will allow us to update our watermark if we need to and always have
it up to date for both scenarios (online and offline).

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- Serve a local watermark in offline scenarios. Upload watermark to the
cnd during packaging.