Ghost image tweaks (#4048)
There's a few things in here:
1. Animation changes
- The shimmer is less sharp, we use a nicely eased gradient instead of a
linear gradient with sharp edges
- When the animation ends, the shimmer fades out instead of disappearing
with a snap
2. Loading bug fix
- When an image first mounts, we currently wait 500ms before showing it.
This is most obvious when duplicating an existing image
- This is the same bug as fixed in #4045, although i think there's extra
stuff in there? Not sure
3. React image preloading `key` trick
- Instead of preloading in an effect with a temporary `Image`, you can
preload with the actual `<img>` you're planning on using with a `key`.
- When the loading is finished, the new `<img>` will show and the old
will be unmounted
- Sometimes (depending on how caching etc. is configured), the separate
`Image` trick can cause a flash whilst the new image is mounted and
fetched from the cache. This doesn't happen with the react/key trick
4. Object URL resource management experiment
- @mimecuvalo and I spent a while talking about different strategies for
managing preview object URLs after the upload has finished. This is
based on some ideas I had after that conversation
- We use reference counting and a timeout to keep the object URL
retained as long as the upload is happening, or anything is still making
use of it.
- imo this isn't worth the complexity. the 60s timeout (we could set it
to 3-5 minutes if we're really worried) is probably good enough, even
though there are times when it theoretically may result in temporarily
broken images on very poor connections.

This PR has a bunch of ideas and they're all a bit muddled together,
sorry! @mimecuvalo we should chat through what's in here. Roughly:
- I would like to land 1 and maybe 2 (depending on #4045). 
- I like 3, but the `Image` approach is probably good enough too. Could
go either way!
- I don't think we should do 4, I don't think the gain is worth the
complexity

### Change type

- [x] `improvement`

---------

Co-authored-by: Mime Čuvalo <mimecuvalo@gmail.com>