Editor.run tweaks (#4067)
Couple of changes to editor.run, the main one being the order.
Previously, we did `editor.run(fn, opts)`. This sort of makes sense,
because options objects usually come last. This diff switched us to
`editor.run(opts, fn)`, which I think makes more sense for a few
reasons:
1. Callback-comes-last is IMO a stronger convention than
options-come-last
2. When the callback is long, having the options all the way down at the
end and possibly off the screen makes it really hard to actually read
what is happening and understand it. We know this code is it a `run`,
but we don't know if that run is affecting history/locking until the
bottom. having the options first makes things much more readable imo
3. Better tooling support from e.g. formatters. This is related to 1,
but the callback-comes-last convention is known by tools like
prettier/biome and so they format it differently. For us, this means
removing an unneeded layer of indentation, and keeping the `run` call
and its options all on one line.

Other changes:
- fix a bug where forced-ness wouldn't be restored properly after a
`run` function
- interface for run options for nicer docs
- clearer name for `._force`