Example node + bun server (#4173)
This PR adds an app with examples for building a tlsync backend in both
node and bun.

Highlights
- simple vite frontend
- node backend using `fastify` (popular modern alternative to express,
and much better api for websockets)
- bun backend using `itty-router` and `Bun.serve`. Interestingly Bun has
a similar websockets api to the cloudflare hibernatable sockets api,
which required some simplification of the socket argument on
`TLSocketRoom`.
- updated itty-router and got rid of itty-cors. There were some minor
breaking API changes to account for, and a new cors api that allows us
to remove an ugly hack.

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [x] `other`

### Test plan

1. Create a shape...
2.

- [ ] Unit tests
- [ ] End to end tests

### Release notes

- Fixed a bug with…