Make asset.fileSize optional (#4206)
This PR makes the `fileSize` property of `TLImageAsset` and
`TLVideoAsset` optional. I first noticed this when I was updating the
Draw Fast repo, but there are a bunch of cases where we don't know the
file size when we're creating an asset. Instead of setting it to -1
(which is sort of magic), we can leave it off if we didn't know it
already.

### Change type

- [x] `api`

### Test plan

- [x] Unit tests

### Release notes

- Made the `fileSize` property of `TLImageAsset` and `TLVideoAsset`
optional