Force flag should override isLocked in more cases (#4214)
After locking the camera movement programmatically, there are several
editor functions that have an option to 'force' their behaviour even
though the camera is locked. But some of these functions do not check
correctly if the force flag is set. This PR changes their behaviour so
you can force camera movement even when the camera is locked.

### Change type

- [x] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Set the camera options `isLocked` to `true`
2. Call `centerOnPoint`, `resetZoom`, `zoomIn`, `zoomOut`,
`zoomToSelection` or `zoomToBounds` with the `force: true` option flag.
3. See that these calls are performed as if `isLocked` was `false`.

- [x] Unit tests
- [ ] End to end tests

### Release notes

- Fixed the `force` flag not being taken in account after locking the
camera and calling `centerOnPoint`, `resetZoom`, `zoomIn`, `zoomOut`,
`zoomToSelection` or `zoomToBounds`.

---------

Co-authored-by: Steve Ruiz <steveruizok@gmail.com>