Allow custom tools to decide whether they can be lockable or not. (#4208)
Right now we have a hardcoded list of tools that are not lockable. This
adds a field to the `StateNode`, which allows the users to customize
whether their custom tools should be lockable or not.

Resolves #2857 

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- Allows custom tools to control whether they can be lockable or not. By
default they are lockable. You can opt out by overriding `StateNode`'s
`isLockable` field:
```typescript
export class MyCustomTool extends StateNode {

   static override isLockable = false

}
```