Add function property warning. (#4266)
Solves https://github.com/tldraw/tldraw/issues/2799

This isn't perfect. It's a bit hard to distinguish these two cases:

```ts
class StateNode {
   id: string
   onExit?: TLExitEventHandler
}
```
That said as soon as you write an `onExit` implementation you will see
the warning. So I'm not sure if it's worth going further than this?


![image](https://github.com/user-attachments/assets/18ef06b2-b740-4d1a-96d1-3d6cd2e50d6a)

Number of warnings:

![image](https://github.com/user-attachments/assets/5b43256f-543c-420d-99bd-331bc52c35c3)

![image](https://github.com/user-attachments/assets/78c869ef-ebe6-4565-a060-8bfcbe4ae8c0)

Happy to go through them if we decide to go with this.

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [x] `other`

### Release notes

- Adds a lint rule to warn when using function properties. We prefer
using methods, since they are more easily extensible.