text shape: dont make a fixed width unless more intentional drag (#4293)
Small thing I noticed when on the call with Jørgen. He created a narrow
vertical column of text because he accidentally dragged just a touch and
didn't know how to undo that state. This makes that drag state a little
less accident-prone. (Maybe it means or dragging state is a touch too
sensitive in general? I dunno)

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`


- [ ] Unit tests
- [ ] End to end tests

### Release notes

- Text shape: dont make a fixed width unless more intentional drag

---------

Co-authored-by: Steve Ruiz <steveruizok@gmail.com>