[Feature, Example] Text search example and `getText` API (#4306)
Motivated by #3495 

An example of how to add custom search via`⌘+f` keyboard shortcuts. This
also adds `getText` method to `ShapeUtil`, which allows the users to
query the shapes for text.

Down the line we could choose to add something like `getShapesWithText`
to the editor and maybe even add some UI to dotcom. For now, let's just
expose the API.


https://github.com/user-attachments/assets/5631301c-a02a-4b66-8dbf-09571e67a46c

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [x] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- Adds `getText` to the `ShapeUtil` api so that we can allow searching
for text in a nicely extensible way.
- Adds an example of how to add text search.