Preserve focus search param (#4344)
This PR adds a kinda-secret "preserve focus" mode using the
`preserveFocus` search param, along with the behavior for handling focus
in that state.

In many of our examples, we want `autoFocus` to be true. However, in our
docs site, we want each example's `autoFocus` to be false, and for the
editor to be focused only when clicked on and blurred when clicking
away.

One solution, which this PR implements, is to use a search parameter
that says "if the URL has a `preserveFocus` search parameter, then even
if `autoFocus` is true, start unfocused`. It also implements that "click
to focus, click away to blur" behavior.

We could add support via a prop, however I'm not entirely sure about
this feature so would prefer to keep it internal for now. Either way, we
would want our docs examples to all use `preserveFocus` routes.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Visit the examples page in our docs.
2. The example should not have focus. Kbds should not work.
3. Click on the example to give it focus. Kbds should work.
4. Click away to blur the editor. Kbds should not work.