Reference docs fixes (#4351)
Restoring some functionality and making some other tweaks from the
current docs site into the new one.

### 1. Code links
Our reference docs link to other docs pages from code and signature
snippets.


![image](https://github.com/user-attachments/assets/09067bdc-4a83-47f6-a5b2-78fd995b5851)

I'm not 100% on the icon here, but i've borrowed this from the
'inherited from' thing:

![image](https://github.com/user-attachments/assets/194b2d41-f7ae-4987-a530-c596c99a4aee)

I think this looks good on the 'inherited from', but i'm less sure about
it inline in the code. The arrow looks a little bit too much like text,
i think? V open to suggestions here. I think we do need some kind of an
icon here to indicate that these are clickable though.

### 2. Focus lines

We highlight key changed lines in certain places:


![image](https://github.com/user-attachments/assets/594f9e57-d47f-46db-9639-f5c8dbcbede1)

Very open to suggestions on the specifics on how this looks.

### 3. Smart links in MDX content

In the old docs, we have a smart `<a>` tag that:
- Uses next's `<Link />` for internal navigation so we get fast
navigation between docs pages
- Sets `target="_blank"` on external links so they open in a new tab

In the new docs, we use plain `<a>` tags that give normal page loads
between each page which are slower. This restores the old link
behaviour.

### 4. API entry "tag" alignment


![image](https://github.com/user-attachments/assets/63add549-0e83-4325-a6ed-52a27a98b36a)

I felt like the vertical alignment of these tags on the new design was a
bit off - they feel a little too high up for me. I tried nudging them
down so they were more centred on the text, but then the baselines being
off felt wrong. in the end I gave up and moved them back over to the
right. Happy to back this change out or adjust further though, but to me
these don't feel quite right where they are currently.

### Change type

- [x] `other`

Co-authored-by: Steve Ruiz <steveruizok@gmail.com>